Enable partial entries to match scaling appetite
