+++ AGENTS.md
@@ -17,13 +17,18 @@
 Overlay policy (baseline bất biến)
 - Baseline: `config/policy_default.json` không bị workflow ghi đè.
 - Nightly overlay: `config/policy_nightly_overrides.json` do calibrators sinh; commit riêng.
-- AI overlay: `config/policy_ai_overrides.json` do tuner sinh; guardrails giới hạn bề mặt.
+- AI overlay: `config/policy_ai_overrides.json` do tuner sinh; ghi đè trực tiếp overlay AI hiện tại.
 - Runtime merge: engine hợp nhất baseline → nightly → ai → legacy (nếu có) thành `out/orders/policy_overrides.json` và dùng cho phiên chạy.
 
 ## Phân chia tài liệu (Docs structure)
 - `README.md`: chỉ tập trung vào cách sử dụng (install, chạy nhanh, I/O, lệnh tiện ích, server usage cơ bản, FAQ). Không chứa chi tiết kỹ thuật/thuật toán; nếu cần, đặt link trỏ sang SYSTEM_DESIGN.md.
-- `SYSTEM_DESIGN.md`: kiến trúc, pipeline, thuật toán/quy tắc ra quyết định, calibrations, execution diagnostics, policy/guardrails, lộ trình tích hợp (API, async runner). Đây là nơi chứa toàn bộ chi tiết kỹ thuật.
+- `SYSTEM_DESIGN.md`: kiến trúc, pipeline, thuật toán/quy tắc ra quyết định, calibrations, execution diagnostics, policy overlays, lộ trình tích hợp (API, async runner). Đây là nơi chứa toàn bộ chi tiết kỹ thuật.
 - `AGENTS.md`: Way of Work, tiêu chuẩn CI/Fail‑fast, quy tắc dữ liệu & contract, convention khi đóng góp. Đồng thời duy trì quy ước phân chia tài liệu này.
+
+## Khẩu vị đầu tư
+- Ưu tiên chốt lời và cắt lỗ sớm, có thể thực hiện theo từng phần để khóa lợi nhuận hoặc hạn chế rủi ro dần.
+- Khi điều kiện thuận lợi, giải ngân bằng các lệnh mua nhỏ, tránh dồn vào một giao dịch quá lớn.
+- Không chờ “thời điểm hoàn hảo” mới mua khối lượng lớn; thay vào đó phân bổ giao dịch theo nhịp thị trường.
 
 Policy biên tập:
 - Khi thêm chi tiết kỹ thuật mới, cập nhật `SYSTEM_DESIGN.md` và chỉ bổ sung gạch đầu dòng/link tối thiểu ở `README.md` nếu ảnh hưởng tới cách dùng.
