+++ tests/test_generate_policy_fileflow.py
@@ -7,7 +7,6 @@ from types import SimpleNamespace
 from unittest.mock import patch
 
 import scripts.ai.codex_policy_budget_bias_tuner as budget_tuner
-import scripts.ai.guardrails as guardrails
 
 
 class TestGeneratePolicyFileflow(unittest.TestCase):
@@ -16,7 +15,7 @@ class TestGeneratePolicyFileflow(unittest.TestCase):
         (base / 'config').mkdir(parents=True, exist_ok=True)
         (base / 'out' / 'debug').mkdir(parents=True, exist_ok=True)
         (base / 'config' / 'policy_overrides.sample.json').write_text('{"schema":"minimal"}', encoding='utf-8')
-        # Baseline policy required for guardrails
+        # Baseline policy to provide schema context
         baseline = {
             "buy_budget_frac": 0.10,
             "add_max": 4,
@@ -46,13 +45,6 @@ class TestGeneratePolicyFileflow(unittest.TestCase):
             budget_tuner.BASE_DIR = base
             budget_tuner.CONFIG_DIR = base / 'config'
             budget_tuner.OUT_DIR = base / 'out'
-            guardrails.BASE_DIR = base
-            guardrails.CONFIG_DIR = base / 'config'
-            guardrails.OUT_DIR = base / 'out'
-            guardrails.STATE_PATH = guardrails.OUT_DIR / 'orders' / 'ai_override_state.json'
-            guardrails.DEFAULT_POLICY_PATH = guardrails.CONFIG_DIR / 'policy_default.json'
-            guardrails.METRICS_PATH = guardrails.OUT_DIR / 'metrics.csv'
-
             def fake_run(cmd, input=None, stdout=None, stderr=None, check=None, cwd=None):
                 # Simulate model writing analysis and output JSON file in cwd
                 Path(cwd, budget_tuner.ANALYSIS_FILENAME).write_text('round analysis', encoding='utf-8')
@@ -65,7 +57,7 @@ class TestGeneratePolicyFileflow(unittest.TestCase):
                 budget_tuner.main()
 
             # Expect copied file exists with correct content
-                dest = base / 'config' / 'policy_ai_overrides.json'
+            dest = base / 'config' / 'policy_ai_overrides.json'
             self.assertTrue(dest.exists())
             self.assertEqual(json.loads(dest.read_text(encoding='utf-8')), {"buy_budget_frac": 0.12})
 
@@ -76,13 +68,6 @@ class TestGeneratePolicyFileflow(unittest.TestCase):
             budget_tuner.BASE_DIR = base
             budget_tuner.CONFIG_DIR = base / 'config'
             budget_tuner.OUT_DIR = base / 'out'
-            guardrails.BASE_DIR = base
-            guardrails.CONFIG_DIR = base / 'config'
-            guardrails.OUT_DIR = base / 'out'
-            guardrails.STATE_PATH = guardrails.OUT_DIR / 'orders' / 'ai_override_state.json'
-            guardrails.DEFAULT_POLICY_PATH = guardrails.CONFIG_DIR / 'policy_default.json'
-            guardrails.METRICS_PATH = guardrails.OUT_DIR / 'metrics.csv'
-
             def fake_run(cmd, input=None, stdout=None, stderr=None, check=None, cwd=None):
                 Path(cwd, budget_tuner.ANALYSIS_FILENAME).write_text('some analysis', encoding='utf-8')
                 # Do NOT write output file here
@@ -102,13 +87,6 @@ class TestGeneratePolicyFileflow(unittest.TestCase):
             budget_tuner.BASE_DIR = base
             budget_tuner.CONFIG_DIR = base / 'config'
             budget_tuner.OUT_DIR = base / 'out'
-            guardrails.BASE_DIR = base
-            guardrails.CONFIG_DIR = base / 'config'
-            guardrails.OUT_DIR = base / 'out'
-            guardrails.STATE_PATH = guardrails.OUT_DIR / 'orders' / 'ai_override_state.json'
-            guardrails.DEFAULT_POLICY_PATH = guardrails.CONFIG_DIR / 'policy_default.json'
-            guardrails.METRICS_PATH = guardrails.OUT_DIR / 'metrics.csv'
-
             def fake_run(cmd, input=None, stdout=None, stderr=None, check=None, cwd=None):
                 Path(cwd, budget_tuner.ANALYSIS_FILENAME).write_text('first round analysis', encoding='utf-8')
                 return SimpleNamespace(stdout=b"CONTINUE\n")
