+++ tests/test_integration_tune_and_orders.py
@@ -10,7 +10,6 @@ from unittest.mock import patch
 import pandas as pd
 
 import scripts.ai.codex_policy_budget_bias_tuner as budget_tuner
-import scripts.ai.guardrails as guardrails
 import scripts.engine.config_io as config_io
 import scripts.engine.pipeline as pipeline
 import scripts.order_engine as oe
@@ -65,14 +64,6 @@ class TestIntegrationTuneAndOrders(unittest.TestCase):
                 stack.enter_context(patch.object(budget_tuner, 'BASE_DIR', base))
                 stack.enter_context(patch.object(budget_tuner, 'CONFIG_DIR', base / 'config'))
                 stack.enter_context(patch.object(budget_tuner, 'OUT_DIR', base / 'out'))
-                stack.enter_context(patch.object(guardrails, 'BASE_DIR', base))
-                stack.enter_context(patch.object(guardrails, 'CONFIG_DIR', base / 'config'))
-                stack.enter_context(patch.object(guardrails, 'OUT_DIR', base / 'out'))
-                stack.enter_context(patch.object(guardrails, 'STATE_PATH', base / 'out' / 'orders' / 'ai_override_state.json'))
-                stack.enter_context(patch.object(guardrails, 'DEFAULT_POLICY_PATH', base / 'config' / 'policy_default.json'))
-                stack.enter_context(patch.object(guardrails, 'METRICS_PATH', base / 'out' / 'metrics.csv'))
-                stack.enter_context(patch.object(guardrails, 'AUDIT_CSV_PATH', base / 'out' / 'orders' / 'ai_overrides_audit.csv'))
-                stack.enter_context(patch.object(guardrails, 'AUDIT_JSONL_PATH', base / 'out' / 'orders' / 'ai_overrides_audit.jsonl'))
                 stack.enter_context(patch('subprocess.run', fake_run))
 
                 budget_tuner.main()
@@ -81,11 +72,7 @@ class TestIntegrationTuneAndOrders(unittest.TestCase):
             self.assertTrue(generated.exists(), 'policy_ai_overrides.json should be written in test mode')
             payload = json.loads(generated.read_text(encoding='utf-8'))
             self.assertIn('buy_budget_frac', payload)
-            self.assertNotIn('news_risk_tilt', payload, 'ephemeral inputs must be stripped by guardrails')
-
-            # Guardrails should persist audit artifacts in test mode as well
-            self.assertTrue((base / 'out' / 'orders' / 'ai_override_state.json').exists())
-            self.assertTrue((base / 'out' / 'orders' / 'ai_overrides_audit.jsonl').exists())
+            self.assertIn('news_risk_tilt', payload, 'raw payload should be preserved without guardrails')
 
     def test_generate_orders_in_test_mode(self) -> None:
         with TemporaryDirectory() as tmp_dir:
